Fix being able to drag a FloatingWindow when dragging its body
We should only be able to drag via title bars, not widget
background.

This separates the concept of Draggable and Window, previously
they were mixed a bit. The draggable is something you drag,
usually a TitleBar, a Tab, or if using native window decorations
then FloatingWindow is also a draggable. Then, for each draggable
there's the actual window being dragged, that's always a FloatingWindow.

We should grab the mouse of the draggable.