Fix anchors out of bounds when redistributing space
In the above case:

+--------+---+
| 1      | 6 |
|        |   |
+---+----|   |
| 2 | 8  |   |
|   |    |   |
+---+----+---+

Supposing 8 is hidden (it's a placeholder), then the separator between 2
and 8 is following the big vertical one between 1 and 6. Now, if the size constraints
of the widget hosted by 2 increases, which might lead to the main window having to increase,
then we have to move that big separator. Problem was it was only honouring the min size of widget 1
because 8 was hidden.