AutoMoc only if needed
The Action.h header contains a couple of includes plus a `Q_OBJECT`
macro inside a block dependent on `!defined(QT_WIDGETS_LIB)`. Running
automoc on that file if `QT_WIDGETS_LIB` is defined causes a warning:

    AutoMoc: path/to/src/Action.h:0:1: note: No relevant classes found. No output generated.

Disable automoc'ing this file if we link against `Qt[56]::Widgets` to
get rid of the warning.

Signed-off-by: Olaf Mandel <o.mandel@menlosystems.com>