tests: Remove tests_launcher
It was a way to speed up test execution, but at the expense of
being weird and non-standard. It would block tst_docks from appearing
as qttest with children, as tst_docks wasn't passed to add_test.

Solution will be to instead split tst_docks into smaller executables,
which was done in 2.0 branch already.