flutter: Introduce C++ coroutines to run tests
Some tests do a sleep, or wait for resize, or wait for window
activation. In Qt, that works well because Qt event loop keeps
running. In flutter however, we need the C++ to return control to
the dart event loop, so we can process events and then return to
where the test left off.

This is easily achievable with C++ coroutines. A test run now looks
something like:

- C++ main() starts
- Creates a Flutter engine which loads main.dart
- main.dart calls a C++ callback that runs the tests
  - all the tests are run with co_await
  - if a test calls, for example co_await wait(1000), then
  all the C++ stack unwinds into dart again, which does normal dart
  event loop processing. Dart pings C++ from time to time and asks
  it if it has coroutines to resume, if yes, then C++ restores
  the stack and continues running the test.

The advantage is we don't have to rewrite 200 tests in Dart.

A disadvantage is that all calls leading to the wait() need to return
an awaiter task. Most testing frameworks are a black box and don't give
us control to changing return types all the way into our tests.
For this reason, need to port away from doctest. For now, my testing
"framework" is simply using Q_ASSERT. To be improved.

Only tst_docks_slow1.cpp has been ported to coroutines, the others will follow.