flutter: Fix some flutter widgets not knowing about their children
Whenever KDDW reparents stuff, it tells flutter about it. However,
if parent is set in the view ctor, there's no flutter counter part
at this point, so flutter would never be notified.

Fix this case by explicitly adding the child to the parent's list
of children, in the ctor case