Allow to create standalone TitleBars
Standalone means it's a regular QWidget without any support for
DND, docking and float button. Just a title and a close button.
No auto-hide either.

One use case is for example to give a title bar to QMessageBox
dialogs under EGLFS, as some users might desire the style to be
shared between the dockwidgets titlebars.