tests: Use logical dpi 96 when running windows/linux tests
Otherwise some tests might fail depending on which system settings
are enabled.

These failures are just off by 1 or 2 pixels when checking if minimum
sizes are honoured, due to QMainWindow applying logical factor on its margins.

Won't have much impact, as tests are run on offscreen QPA anyway, which
is 96dpi, but I want to be able to run the tests on XCB without
unimportant failures