flutter: Developer-mode is now Linux only
You can still build KDDW, but the unit-tests will be missing.

You can still develop KDDW itself with Windows/macOS, but you'll need
to check KDAB CI for the results of the test run.

The tests are a bit involved, as they require custom flutter engines,
and embedders. Reducing maintenance burden this way.