Fill drag pixmap with transparent color
QPixmap has uninitialized data when you first create it, even in the
alpha channel. And the opacity setting on QPainter is additive, it
happens per draw operation. We end up trying to draw images on top of a
bunch of garbage data, so let's clear it with a transparent color first.