Keep track of override cursor in WidgetResizeHandler
QGuiApplication keeps a stack of override cursor and the number of
setOverrideCursor calls must match the number of restoreOverrideCursor
calls. Keep track of whether we've already set the override cursor in
WidgetResizeHandler so we don't end up with the wrong cursor when
restoring.

(cherry picked from commit a96003304ab3a2619226b226f882341dfa736903)