tests: Introduce a layout fuzzer
It generates a random layout from the schema.
It's still not very good at making KDDW crash, as KDDW exits
early due to main window names not existing.

To improve, we need to generate more correct layouts. We already
follow the schema, but we should follow some rules that can't
be written in the schema, like having a finite set of dock widget
names and reference those.

In other words, the generated layouts should not deviate too much from
the rules, but just a bit. If too much then it's easily caught
by current KDDW validation.