Print error when docking into a group being deleted
Not easily reproduced by the user, but can happen programatically
when doing dw->close(), followed by a dw->open().

The close() schedules a Group deletion, while dw->open() restores
it to the group. The group is then deleted on the next event loop
nd deletes the dock widgets.

For now, just printing an error, so tests fail whenever this happen.

Relates to #326.