Remove md linter precommit hook
It has a big amount of false positives. Often complains about
things that render perfectly.

Removing it since it's always getting in the way.
I've written a bunch of documentation in the form of mdbook so can't
waste time fixing things that don't affect rendering.

Next will write a github action that generates static .HTML
from the .md anyway.