Remove KDDOCKWIDGETS_SUPPORTS_NESTED_MAINWINDOWS
It's only declared on the header, not enforced anywhere in implementation.

There are however (unfixable) problems when using 2 main windows, where,
depending on the window manager, the dock being dragged doesn't appear
on top of it's destination main window, because it's parented to the
source main window. Will leave that for the users to experiment with
removing Qt::Tool or setting transient parents, depending on their
window manager needs, which we can't control on Linux.

Fixes #461