Add Config::InternalFlag_NoDeleteLaterWorkaround
We're not using QObject::deleteLater() by default due to QTBUG-83030,
but the downside is that our QTimer leaks at shutdown.

With this new flag, we disable the workaround, as the conditions
in that bug report are pretty rare, and bug is fixed in Qt 6.7
anyway.

The default is still to avoid deleteLater(), as it's much
easier to catch a leak than to catch a posted deferred delete that
never arrived