Raise separators after creating them
Usually unneeded as separators don't overlap with
dock widgets. But we support the use case of having
them overlap, if thickness is increased.

Actual problem is on windows, when using native
child windows (each has a window), and the titlebar
window is eating our separator events, which is
a Qt bug.