Fix case where preferredSize wasn't being honoured
If a container is vertical, we would only honour the preferred
item's height when inserting it. Buf if container wasn't visible
we can honour the preferred width as well. If the container already
had visible items then we don't change the width as that's weird.
Doing dnd and dropping only affects the main-axis length, not
cross-axis.

And vice-versa for horizontal container.