Fix layout being bigger than DropArea in some cases
This could happen when restoring a layout that contained a sidebar.
When the sidebar is restored, the DropArea needs to increase in
size to accomodate it, but we were ignoring the resize because
we were restoring the layout.

Solution is to restore the sidebar before starting to restore
the layout