qtquick: Fix restored windows being invisible
On Windows, window thumbnails will only appear on
alt-tab if windows were rendered at least once. When restoring directly
to minimized state they're not rendered. We use a workaround where
we show the window "normal", with opacity 0, so it's rendered, then
immediately minimize.

The above had a bug, as the floating window flag was being saved to json,
then it would be read later even if the window wasn't minimized. Resulting
in opacity never being restored.