flutter: Rewrite flutter frontend to use only the KDDW layouting
This greatly simplifies the flutter port. KDDW now only drives
the layouting and doesn't abstract tabs, windows and DnD anymore,
which will now be done in Dart.

The amount of over-engineered abstractions in C++ was causing it to be
a pain to even do simple bug fixes. The code is now very clean in dart,
without needing to go back and forth to C++ land for every mouse click.

Allows to get rid of dartagnan which broke whenever shiboken or LLVM
was updated.

Allows to get rid of qcoro co-routines as tests are no longer driven
by C++, no need to mix 2 event loops.

The state of the port is still dependent on flutter supporting
multi-window, but we're now in a position to move faster, as tons
of tech debt have been removed.