Fix restoring maximized window corner case
For issue #553

On Windows it's possible to have normal geometry
in a different screen than the maximized window if
you drag a window across monitors and touch an
aerosnap edge.

While restoring, we need a fake setGeometry()
call to set the normal geometry, since Qt doesn't
provide QWindow::setNormalGeometry(), that resulted
in the window being maximized on the wrong window.

The solution is to patch the normal geometry to be
in the first screen.